+++ assets/javascripts/kitten/components/graphics/icons-next/icons-next.stories.js
@@ -20,6 +20,7 @@ import {
   UsersIconNext,
   WhatsappIconNext,
   WindowEditIconNext,
+  StatusIconNext,
 } from 'kitten'
 import { Text, Title, pxToRem } from 'kitten'
 import React from 'react'
@@ -136,6 +137,23 @@ export const AllIcons = ({ colorInput }) => {
         </IconContainer>
         <IconContainer>
           <WindowEditIconNext color={colorInput} />
+        </IconContainer>
+      </Group>
+      <Group title="Status icons">
+        <IconContainer suffix="status=info">
+          <StatusIconNext status="info" color={colorInput} />
+        </IconContainer>
+        <IconContainer suffix="status=warning">
+          <StatusIconNext status="warning" color={colorInput} />
+        </IconContainer>
+        <IconContainer suffix="status=disabled">
+          <StatusIconNext status="disabled" color={colorInput} />
+        </IconContainer>
+        <IconContainer suffix="status=success">
+          <StatusIconNext status="success" color={colorInput} />
+        </IconContainer>
+        <IconContainer suffix="status=danger">
+          <StatusIconNext status="danger" color={colorInput} />
         </IconContainer>
       </Group>
       <Group title="Social">